refactors to notifications system
 - Renamed several functions and variables for clarity
 - Changed the NotificationStruct class definition to a namedtuple definition (faster reads; more structlike behavior)
 - Moved all systems to a separate module
 - Replaced elif chain with match statement
 - Replaced some if/elses with early return to reduce nesting (improved readability)
 - migrated independent commands to separate function definitions
